changed general artifact to what is not impedence

+++ Detect_Artifacts.m
@@ -140,6 +140,6 @@ if nargout > 1
             end
         end
     end
-    artsStruct.general = ~artsStruct.impedance;
+    artsStruct.general = double(~artsStruct.impedance);
 
 end
